refactored sampler and pooler
