--- src/autora/experimentalist/sampler/falsification/__init__.py
@@ -1,237 +1,0 @@
-from typing import Optional, Tuple, cast, Iterable
-
-import numpy as np
-import torch
-from sklearn.preprocessing import StandardScaler
-from torch import nn
-from torch.autograd import Variable
-
-from autora.experimentalist.pooler.falsification import (
-    class_to_onehot,
-    get_iv_limits,
-    train_popper_net
-)
-from autora.variable import ValueType, VariableCollection
-
-from autora.utils.deprecation import deprecated_alias
-
-
-def falsification_sample(
-    condition_pool,
-    model,
-    reference_conditions: np.ndarray,
-    reference_observations: np.ndarray,
-    metadata: VariableCollection,
-    num_samples: Optional[int] = None,
-    training_epochs: int = 1000,
-    training_lr: float = 1e-3,
-    plot: bool = False,
-):
-    """
-    A Sampler that generates samples of experimental conditions with the objective of maximizing the
-    (approximated) loss of a model relating experimental conditions to observations. The samples are generated by first
-    training a neural network to approximate the loss of a model for all patterns in the training data.
-    Once trained, the network is then provided with the candidate samples of experimental conditions and the selects
-    those with the highest loss.
-
-    Args:
-        condition_pool: The candidate samples of experimental conditions to be evaluated.
-        model: Scikit-learn model, could be either a classification or regression model
-        reference_conditions: Experimental conditions that the model was trained on
-        reference_observations: Observations that the model was trained to predict
-        metadata: Meta-data about the dependent and independent variables specifying the experimental conditions
-        num_samples: Number of samples to return
-        training_epochs: Number of epochs to train the popper network for approximating the
-        error of the model
-        training_lr: Learning rate for training the popper network
-        plot: Print out the prediction of the popper network as well as its training loss
-
-    Returns: Samples with the highest loss
-
-    """
-
-    # format input
-
-    reference_conditions = np.array(reference_conditions)
-    if len(reference_conditions.shape) == 1:
-        reference_conditions = reference_conditions.reshape(-1, 1)
-
-    # get target pattern for popper net
-    model_predict = getattr(model, "predict_proba", None)
-    if callable(model_predict) is False:
-        model_predict = getattr(model, "predict", None)
-
-    if callable(model_predict) is False or model_predict is None:
-        raise Exception("Model must have `predict` or `predict_proba` method.")
-
-    predicted_observations = model_predict(reference_conditions)
-    if isinstance(predicted_observations, np.ndarray) is False:
-        try:
-            predicted_observations = np.array(predicted_observations)
-        except Exception:
-            raise Exception("Model prediction must be convertable to numpy array.")
-    if predicted_observations.ndim == 1:
-        predicted_observations = predicted_observations.reshape(-1, 1)
-
-    new_conditions, scores = falsification_score_sampler_from_predictions(
-        condition_pool,
-        predicted_observations,
-        reference_conditions,
-        reference_observations,
-        metadata,
-        num_samples,
-        training_epochs,
-        training_lr,
-        plot,
-    )
-
-    return new_conditions
-
-def falsification_score_sample(
-    condition_pool,
-    model,
-    reference_conditions: np.ndarray,
-    reference_observations: np.ndarray,
-    metadata: Optional[VariableCollection] = None,
-    num_samples: Optional[int] = None,
-    training_epochs: int = 1000,
-    training_lr: float = 1e-3,
-    plot: bool = False,
-):
-    """
-    A Sampler that generates samples of experimental conditions with the objective of maximizing the
-    (approximated) loss of a model relating experimental conditions to observations. The samples are generated by first
-    training a neural network to approximate the loss of a model for all patterns in the training data.
-    Once trained, the network is then provided with the candidate samples of experimental conditions and the selects
-    those with the highest loss.
-
-    Args:
-        condition_pool: The candidate samples of experimental conditions to be evaluated.
-        model: Scikit-learn model, could be either a classification or regression model
-        reference_conditions: Experimental conditions that the model was trained on
-        reference_observations: Observations that the model was trained to predict
-        metadata: Meta-data about the dependent and independent variables specifying the experimental conditions
-        num_samples: Number of samples to return
-        training_epochs: Number of epochs to train the popper network for approximating the
-        error of the model
-        training_lr: Learning rate for training the popper network
-        plot: Print out the prediction of the popper network as well as its training loss
-
-    Returns:
-        new_conditions: Samples of experimental conditions with the highest loss
-        scores: Normalized falsification scores for the samples
-
-    """
-
-    if isinstance(reference_conditions, Iterable):
-        reference_conditions = np.array(list(reference_conditions))
-
-    reference_conditions = np.array(reference_conditions)
-    if len(reference_conditions.shape) == 1:
-        reference_conditions = reference_conditions.reshape(-1, 1)
-
-    predicted_observations = model.predict(reference_conditions)
-
-    return falsification_score_sample_from_predictions(condition_pool,
-                                                        predicted_observations,
-                                                        reference_conditions,
-                                                        reference_observations,
-                                                        metadata,
-                                                        num_samples,
-                                                        training_epochs,
-                                                        training_lr,
-                                                        plot)
-
-
-def falsification_score_sample_from_predictions(
-    condition_pool,
-    predicted_observations: np.ndarray,
-    reference_conditions: np.ndarray,
-    reference_observations: np.ndarray,
-    metadata: Optional[VariableCollection] = None,
-    num_samples: Optional[int] = None,
-    training_epochs: int = 1000,
-    training_lr: float = 1e-3,
-    plot: bool = False,
-):
-    """
-    A Sampler that generates samples of experimental conditions with the objective of maximizing the
-    (approximated) loss of a model relating experimental conditions to observations. The samples are generated by first
-    training a neural network to approximate the loss of a model for all patterns in the training data.
-    Once trained, the network is then provided with the candidate samples of experimental conditions and the selects
-    those with the highest loss.
-
-    Args:
-        condition_pool: The candidate samples of experimental conditions to be evaluated.
-        predicted_observations: Prediction obtained from the model for the set of reference experimental conditions
-        reference_conditions: Experimental conditions that the model was trained on
-        reference_observations: Observations that the model was trained to predict
-        metadata: Meta-data about the dependent and independent variables specifying the experimental conditions
-        num_samples: Number of samples to return
-        training_epochs: Number of epochs to train the popper network for approximating the
-        error of the model
-        training_lr: Learning rate for training the popper network
-        plot: Print out the prediction of the popper network as well as its training loss
-
-    Returns:
-        new_conditions: Samples of experimental conditions with the highest loss
-        scores: Normalized falsification scores for the samples
-
-    """
-
-    if isinstance(condition_pool, Iterable):
-        condition_pool = np.array(list(condition_pool))
-
-    if isinstance(condition_pool, list):
-        condition_pool = np.array(condition_pool)
-
-    if isinstance(condition_pool, np.ndarray) is False:
-        raise Exception("condition_pool must be a numpy array.")
-
-    if len(condition_pool.shape) == 1:
-        condition_pool = condition_pool.reshape(-1, 1)
-
-    reference_conditions = np.array(reference_conditions)
-    if len(reference_conditions.shape) == 1:
-        reference_conditions = reference_conditions.reshape(-1, 1)
-
-    reference_observations = np.array(reference_observations)
-    if len(reference_observations.shape) == 1:
-        reference_observations = reference_observations.reshape(-1, 1)
-
-    if num_samples is None:
-        num_samples = condition_pool.shape[0]
-
-    if metadata is not None:
-        if metadata.dependent_variables[0].type == ValueType.CLASS:
-            # find all unique values in reference_observations
-            num_classes = len(np.unique(reference_observations))
-            reference_observations = class_to_onehot(reference_observations, n_classes=num_classes)
-
-    # create list of IV limits
-    iv_limit_list = get_iv_limits(reference_conditions, metadata)
-
-    popper_net, model_loss = train_popper_net(predicted_observations,
-                                              reference_conditions,
-                                              reference_observations,
-                                              metadata,
-                                              iv_limit_list,
-                                              training_epochs,
-                                              training_lr,
-                                              plot)
-
-    # now that the popper network is trained we can assign losses to all data points to be evaluated
-    popper_input = Variable(torch.from_numpy(condition_pool)).float()
-    Y = popper_net(popper_input).detach().numpy().flatten()
-    scaler = StandardScaler()
-    score = scaler.fit_transform(Y.reshape(-1, 1)).flatten()
-
-    # order rows in Y from highest to lowest
-    sorted_conditions = condition_pool[np.argsort(score)[::-1]]
-    sorted_score = score[np.argsort(score)[::-1]]
-
-    return sorted_conditions[:num_samples], sorted_score[:num_samples]
-
-falsification_sampler = deprecated_alias(falsification_sample, "falsification_sampler")
-falsification_score_sampler = deprecated_alias(falsification_score_sample, "falsification_score_sampler")
-falsification_score_sampler_from_predictions = deprecated_alias(falsification_score_sample_from_predictions, "falsification_score_sampler_from_predictions")
