+++ tests/test_exp_falsification_pooler.py
@@ -1,9 +1,10 @@
 import numpy as np
+import pandas as pd
 import pytest
 import torch
 from sklearn.linear_model import LinearRegression, LogisticRegression
 
-from autora.experimentalist.pooler.falsification import falsification_pool
+from autora.experimentalist.falsification import falsification_pool
 from autora.variable import DV, IV, ValueType, VariableCollection
 
 
@@ -157,6 +158,71 @@ def test_falsification_pool_regression(synthetic_linr_model, seed):
         assert condition < 0.1 or condition > 6.1 or \
                (condition < 2.5 and condition > 1.5)  or \
                (condition < 5 and condition > 4)
+
+def test_falsification_pandas(
+    synthetic_logr_model, seed
+):
+
+    # Import model and data_closed_loop
+    conditions, observations = get_xor_data()
+    model = synthetic_logr_model
+
+    conditions = pd.DataFrame(conditions, columns=["x1", "x2"])
+    observations = pd.DataFrame(observations, columns=["y"])
+
+    # Specify independent variables
+    iv1 = IV(
+        name="x1",
+        value_range=(0, 1),
+        units="intensity",
+        variable_label="stimulus 1",
+    )
+
+    iv2 = IV(
+        name="x2",
+        value_range=(0, 1),
+        units="intensity",
+        variable_label="stimulus 2",
+    )
+
+    # specify dependent variables
+    dv1 = DV(
+        name="y",
+        value_range=(0, 1),
+        units="class",
+        variable_label="class",
+        type=ValueType.CLASS,
+    )
+
+    # Variable collection with ivs and dvs
+    metadata = VariableCollection(
+        independent_variables=[iv1, iv2],
+        dependent_variables=[dv1],
+    )
+
+    # Run falsification pooler
+    new_conditions = falsification_pool(
+        model=model,
+        reference_conditions=conditions,
+        reference_observations=observations,
+        metadata=metadata,
+        num_samples=2,
+        training_epochs=1000,
+        optimization_epochs=1000,
+        training_lr=1e-3,
+        optimization_lr=1e-3,
+        limit_offset=10 ** -10,
+        limit_repulsion=0,
+        plot=False
+    )
+
+    # convert Iterable to numpy array
+    new_conditions = np.array(list(new_conditions))
+
+    # Check that at least one of the resulting samples is the one that is
+    # underrepresented in the data_closed_loop used for model training
+    assert (new_conditions[0,0] > 0.99 and new_conditions [0,1] > 0.99) or \
+           (new_conditions[1,0] > 0.99 and new_conditions [1,1] > 0.99)
 
 def test_doc_example():
     # Specify X and Y
